requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.8
djangorestframework==3.12.2
flake8==3.8.4
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
iniconfig==1.1.1
mccabe==0.6.1
msgpack==1.0.2
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dotenv==0.17.1
pytz==2020.5
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1; platform_system == "Windows"
txaio==20.12.1
whitenoise==5.2.0
zope.interface==5.2.0
